feat: remove lerna dependency, close #194 (#195)
BREAKING CHANGE: Major rewrite of monodeploy. It is now a wrapper around Yarn Berry's API, and operates directly on yarn workspaces rather than lerna. It is no longer compatible with lerna monorepos. Migration instructions are available in the README. The API and CLI interface have completely changed. Please refer to the documentation for the new usage.