fix: apply version strategy to largest version among group
This fixes a bug in an earlier v3.0.0 release candidate, where it was possible for monodeploy to publish a new version of a group with a value smaller than the latest release. This also addresses issues around non-idiomatic versioning. So a major version strategy is now guaranteed to yield x.0.0, while minor is guaranteed to yield x.x.0. This guarantee regressed in an earlier release candidate, and was incorrectly believed to be correct behaviour. The test coverage covering this functionality has been updated with comments.