feat: run all pre-lifecycle hooks before publish, close #311 (#510)
NOTE: This allows prepack failures to short-circuit the publish pipeline rather than leave the project in a half-published state. Originally, this was going to be behind a opt-in flag, however that would drastically increase complexity of the pipeline. This shouldn't introduce any breaking behaviour. Please open a GitHub issue if it does and we can revert this and re-introduce behind a flag.