feat: add versionStrategy config (#659)
NOTE: Set versionStrategy.coerceImplicitPeerDependency to patch/minor/major to round down the patched peerDependency version prior to publishing. Previously, if Pkg-A declared a peer on Pkg-B, the peerDependency version range would be kept up to date with the latest version of Pkg-B. This may be undesirable as it can produce unnecessary peer dependency incompatibility warnings when the public API of Pkg-B did not change from patch to patch.